+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.5.0(2020-01-12)
+
+* Bump dependencies
+* Add `after_request`
+* Add flexible routing strategy
+
 ## v0.4.0(2019-12-16)
 
 * Add `max_upload_bytes`
@@ -12,7 +18,7 @@
 ## v0.2.0(2019-11-25)
 
 * Fix typo
-* Update dependencies
+* Bump dependencies
 
 ## v0.1.0(2019-11-17)